backend/requirements.txt
fastapi==0.114.2
certifi==2024.8.30 
uvicorn==0.30.6
asgi-ratelimit[full]
scipy==1.14.0
numpy==1.26.4
transformers==4.44.2
langchain_huggingface==0.0.3
langchain-qdrant==0.1.3
FlagEmbedding==1.2.11
peft==0.12.0
python-dotenv==1.0.1